ai_service/requirements.txt
grpcio>=1.60.0
grpcio-tools>=1.60.0
protobuf>=4.25.0
ultralytics>=8.1.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
torch>=2.1.0
